Proposition for non-conversions and maybe bug
Thanks for the great help. 
This change would allow not converting a value in case you just want to graph it. I just wanted to visualise values and it kept closing the program with no case for 'nothing'
Also please test if the order of the 4 hormones has to be changed in the list like that. In line 157 you access the element 0 so i think it should be the graphed/displayed one when you create your list in line 11. (?)
I got wrong results when testing, the factor seemed to be using another conversion.
If something is bad, ignore it.
Thanks